Add `jlinkOptions` parameter to pass options to jlink executable
